feat: 모든 미팅방 확인하기 api

+++ src/main/kotlin/com/gcc/miti/module/entity/Group.kt
@@ -19,7 +19,7 @@ import javax.persistence.Table
 class Group(
     val description: String,
     val title: String,
-    val maxUsers: Long,
+    val maxUsers: Int,
 
     @Enumerated
     val groupStatus: GroupStatus,
